scripts/chapter4/pyproject.toml
[project]
name = "chapter4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain==0.3.27",
    "langchain-openai==0.3.33",
    "langchain-community==0.3.29",
    "langgraph==0.6.7",
    "python-dotenv==1.1.1",
    "jupyterlab==4.4.7",
]
